Fixed a bug and various improvements to the code:
Fixed a bug where tags were not generated if a directory
containing a file being parsed for tags contained spaces in the
directory path. More work on the special include option. It still
does not work yet, but the work done on it does not break any
existing functionality and all the tests are passing. The tests
for parts of this new feature that are not finished yet are
commented out so that all the tests will pass. Refactored the
tests. Moved the generated test files and directories outside the
tests directory to make it easier to write a UTRun command that
safely runs all the tests with a single command. Made the tests
for the core feature more thorough. Now they more thoroughly ensure
the expected tag was generated in the tags file. They also make
sure no unwanted tags were generated.