Adding support for more languages can now be done as a list:
Adding languages can still be done the old way, or you can pass in
a list of strings instead of a single string. This allows you the
user (and the internal code) to add multiple languages with a
single function call. An example is already in the code to add
support for PHP and JavaScript out of the box:
call project_tags#add_extension(['php', 'js'])